+++ src/psyclone/psyir/frontend/fparser2.py
@@ -78,6 +78,117 @@ INTENT_MAPPING = {"in": ArgumentInterface.Access.READ,
                   "inout": ArgumentInterface.Access.READWRITE}
 
 
+def _find_or_create_imported_symbol(location, name, scope_limit=None,
+        **kargs):
+    '''Returns the symbol with the name 'name' from a symbol table
+    associated with this node or one of its ancestors.  If the symbol
+    is not found and there are no ContainerSymbols with wildcard imports
+    then an exception is raised. However, if there are one or more
+    ContainerSymbols with wildcard imports (which could therefore be
+    bringing the symbol into scope) then a new Symbol with the
+    specified visibility but of unknown interface is created and
+    inserted in the most local SymbolTable that has such an import.
+    The scope_limit variable further limits the symbol table search so
+    that the search through ancestor nodes stops when the scope_limit node
+    is reached i.e. ancestors of the scope_limit node are not searched.
+
+    :param str name: the name of the symbol.
+    :param scope_limit: optional Node which limits the symbol \
+        search space to the symbol tables of the nodes within the \
+        given scope. If it is None (the default), the whole \
+        scope (all symbol tables in ancestor nodes) is searched \
+        otherwise ancestors of the scope_limit node are not \
+        searched.
+    :type scope_limit: :py:class:`psyclone.psyir.nodes.Node` or \
+        `NoneType`
+    :param visibility: the visibility to give to any new symbol.
+    :type visibility: :py:class:`psyclone.Symbol.Visbility`
+
+    :returns: the matching symbol.
+    :rtype: :py:class:`psyclone.psyir.symbols.Symbol`
+
+    :raises TypeError: if the supplied scope_limit is not a Node.
+    :raises ValueError: if the supplied scope_limit node is not an \
+        ancestor of the supplied node.
+    :raises SymbolError: if no matching symbol is found and there are \
+        no ContainerSymbols from which it might be brought into scope.
+
+    '''
+    if scope_limit:
+        # Validate the supplied scope_limit
+        if not isinstance(scope_limit, Node):
+            raise TypeError(
+                "The scope_limit argument '{0}' provided to the "
+                "find_or_create_symbol method, is not of type `Node`."
+                "".format(str(scope_limit)))
+
+        # Check that the scope_limit Node is an ancestor of this
+        # Reference Node and raise an exception if not.
+        mynode = location.parent
+        while mynode is not None:
+            if mynode is scope_limit:
+                # The scope_limit node is an ancestor of the
+                # supplied node.
+                break
+            mynode = mynode.parent
+        else:
+            # The scope_limit node is not an ancestor of the
+            # supplied node so raise an exception.
+            raise ValueError(
+                "The scope_limit node '{0}' provided to the "
+                "find_or_create_symbol method, is not an ancestor of this "
+                "node '{1}'.".format(str(scope_limit), str(location)))
+
+    # Keep a reference to the most local SymbolTable with a wildcard
+    # import in case we need to create a Symbol.
+    first_symbol_table = None
+    test_node = location
+
+    # Iterate over ancestor Nodes of this Node.
+    while test_node:
+        # For simplicity, test every Node for the existence of a
+        # SymbolTable (rather than checking for the particular
+        # Node types which we know to have SymbolTables).
+        if hasattr(test_node, 'symbol_table'):
+            # This Node does have a SymbolTable.
+            symbol_table = test_node.symbol_table
+
+            try:
+                # If the reference matches a Symbol in this
+                # SymbolTable then return the Symbol.
+                return symbol_table.lookup(name, scope_limit=test_node)
+            except KeyError:
+                # The Reference Node does not match any Symbols in
+                # this SymbolTable. Does this SymbolTable have any
+                # wildcard imports?
+                for csym in symbol_table.containersymbols:
+                    if csym.wildcard_import:
+                        if not first_symbol_table:
+                            first_symbol_table = symbol_table
+                            break
+
+        if test_node is scope_limit:
+            # The ancestor scope/top-level Node has been reached and
+            # nothing has matched.
+            break
+
+        # Move on to the next ancestor.
+        test_node = test_node.parent
+
+    if first_symbol_table:
+        # No symbol found but there are one or more Containers from which
+        # it may be being brought into scope. Therefore create a generic
+        # Symbol with a deferred interface and add it to the most
+        # local SymbolTable with a wildcard import.
+        return first_symbol_table.new_symbol(name,
+                interface=UnresolvedInterface(), **kargs)
+
+    # All requested Nodes have been checked but there has been no
+    # match and there are no wildcard imports so raise an exception.
+    raise SymbolError(
+        "No Symbol found for name '{0}'.".format(name))
+
+
 def _check_args(array, dim):
     '''Utility routine used by the _check_bound_is_full_extent and
     _check_array_range_literal functions to check common arguments.
@@ -1319,7 +1430,7 @@ class Fparser2Reader(object):
             # This is a variable of derived type
             type_name = str(walk(type_spec, Fortran2003.Type_Name)[0])
             # Do we already have a Symbol for this derived type?
-            type_symbol = parent.find_or_create_symbol(type_name)
+            type_symbol = _find_or_create_imported_symbol(parent, type_name)
             # pylint: disable=unidiomatic-typecheck
             if type(type_symbol) == Symbol:
                 # We do but we didn't know what kind of symbol it was. Create
@@ -1698,7 +1809,8 @@ class Fparser2Reader(object):
                         # If a suitable unqualified use statement is found then
                         # this call creates a Symbol and inserts it in the
                         # appropriate symbol table.
-                        parent.find_or_create_symbol(name, visibility=vis)
+                        _find_or_create_imported_symbol(parent, name,
+                                                        visibility=vis)
                     except SymbolError as err:
                         # Improve the error message with context-specific info
                         six.raise_from(SymbolError(
@@ -1974,7 +2086,8 @@ class Fparser2Reader(object):
         loop_var = str(ctrl[0].items[1][0])
         variable_name = str(loop_var)
         try:
-            data_symbol = parent.find_or_create_symbol(variable_name)
+            data_symbol = _find_or_create_imported_symbol(parent,
+                                                          variable_name)
         except SymbolError:
             raise InternalError(
                 "Loop-variable name '{0}' is not declared and there are no "
@@ -2450,7 +2563,8 @@ class Fparser2Reader(object):
             range_idx = 0
             for idx, child in enumerate(array.children):
                 if isinstance(child, Range):
-                    symbol = array.find_or_create_symbol(loop_vars[range_idx])
+                    symbol = _find_or_create_imported_symbol(
+                                array, loop_vars[range_idx])
                     array.children[idx] = Reference(symbol, parent=array)
                     range_idx += 1
 
@@ -2568,7 +2682,8 @@ class Fparser2Reader(object):
             size_node = BinaryOperation(BinaryOperation.Operator.SIZE,
                                         parent=loop)
             loop.addchild(size_node)
-            symbol = size_node.find_or_create_symbol(arrays[0].name)
+            symbol = _find_or_create_imported_symbol(
+                        size_node, arrays[0].name)
 
             size_node.addchild(Reference(symbol, parent=size_node))
             size_node.addchild(Literal(str(idx), integer_type,
@@ -2854,7 +2969,7 @@ class Fparser2Reader(object):
         :rtype: :py:class:`psyclone.psyir.nodes.Reference`
 
         '''
-        symbol = parent.find_or_create_symbol(node.string)
+        symbol = _find_or_create_imported_symbol(parent, node.string)
         return Reference(symbol, parent)
 
     def _parenthesis_handler(self, node, parent):
@@ -2894,7 +3009,7 @@ class Fparser2Reader(object):
 
         '''
         reference_name = node.items[0].string.lower()
-        symbol = parent.find_or_create_symbol(reference_name)
+        symbol = _find_or_create_imported_symbol(parent, reference_name)
 
         array = ArrayReference(symbol, parent)
         self.process_nodes(parent=array, nodes=node.items[1].items)
