+++ src/psyclone/psyir/nodes/node.py
@@ -1138,132 +1138,6 @@ class Node(object):
             "Unable to find the scope of node '{0}' as none of its ancestors "
             "are Container or Schedule nodes.".format(self))
 
-    def find_or_create_symbol(self, name, scope_limit=None,
-                              visibility=Symbol.DEFAULT_VISIBILITY):
-        '''Returns the symbol with the name 'name' from a symbol table
-        associated with this node or one of its ancestors.  If the symbol
-        is not found and there are no ContainerSymbols with wildcard imports
-        then an exception is raised. However, if there are one or more
-        ContainerSymbols with wildcard imports (which could therefore be
-        bringing the symbol into scope) then a new Symbol with the
-        specified visibility but of unknown interface is created and
-        inserted in the most local SymbolTable that has such an import.
-        The scope_limit variable further limits the symbol table search so
-        that the search through ancestor nodes stops when the scope_limit node
-        is reached i.e. ancestors of the scope_limit node are not searched.
-
-        :param str name: the name of the symbol.
-        :param scope_limit: optional Node which limits the symbol \
-            search space to the symbol tables of the nodes within the \
-            given scope. If it is None (the default), the whole \
-            scope (all symbol tables in ancestor nodes) is searched \
-            otherwise ancestors of the scope_limit node are not \
-            searched.
-        :type scope_limit: :py:class:`psyclone.psyir.nodes.Node` or \
-            `NoneType`
-        :param visibility: the visibility to give to any new symbol.
-        :type visibility: :py:class:`psyclone.Symbol.Visbility`
-
-        :returns: the matching symbol.
-        :rtype: :py:class:`psyclone.psyir.symbols.Symbol`
-
-        :raises TypeError: if the supplied scope_limit is not a Node.
-        :raises ValueError: if the supplied scope_limit node is not an \
-            ancestor of the supplied node.
-        :raises TypeError: if the supplied visibility is not of \
-            `Symbol.Visibility` type.
-        :raises SymbolError: if no matching symbol is found and there are \
-            no ContainerSymbols from which it might be brought into scope.
-
-        '''
-        if scope_limit:
-            # Validate the supplied scope_limit
-            if not isinstance(scope_limit, Node):
-                raise TypeError(
-                    "The scope_limit argument '{0}' provided to the "
-                    "find_or_create_symbol method, is not of type `Node`."
-                    "".format(str(scope_limit)))
-
-            # Check that the scope_limit Node is an ancestor of this
-            # Reference Node and raise an exception if not.
-            mynode = self.parent
-            while mynode is not None:
-                if mynode is scope_limit:
-                    # The scope_limit node is an ancestor of the
-                    # supplied node.
-                    break
-                mynode = mynode.parent
-            else:
-                # The scope_limit node is not an ancestor of the
-                # supplied node so raise an exception.
-                raise ValueError(
-                    "The scope_limit node '{0}' provided to the "
-                    "find_or_create_symbol method, is not an ancestor of this "
-                    "node '{1}'.".format(str(scope_limit), str(self)))
-
-        # Validate supplied visibility
-        if not isinstance(visibility, Symbol.Visibility):
-            raise TypeError(
-                "The visibility argument '{0}' provided to the "
-                "find_or_create_symbol method should be of `Symbol."
-                "Visibility` type but instead is: '{1}'."
-                "".format(str(visibility), type(visibility).__name__))
-
-        # Keep a list of (symbol table, container) tuples for containers that
-        # have wildcard imports into the current scope and therefore may
-        # contain the symbol we're searching for.
-        possible_containers = []
-        # Keep a reference to the most local SymbolTable with a wildcard
-        # import in case we need to create a Symbol.
-        first_symbol_table = None
-        test_node = self
-
-        # Iterate over ancestor Nodes of this Node.
-        while test_node:
-            # For simplicity, test every Node for the existence of a
-            # SymbolTable (rather than checking for the particular
-            # Node types which we know to have SymbolTables).
-            if hasattr(test_node, 'symbol_table'):
-                # This Node does have a SymbolTable.
-                symbol_table = test_node.symbol_table
-
-                try:
-                    # If the reference matches a Symbol in this
-                    # SymbolTable then return the Symbol.
-                    return symbol_table.lookup(name, scope_limit=test_node)
-                except KeyError:
-                    # The Reference Node does not match any Symbols in
-                    # this SymbolTable. Does this SymbolTable have any
-                    # wildcard imports?
-                    for csym in symbol_table.containersymbols:
-                        if csym.wildcard_import:
-                            possible_containers.append((symbol_table, csym))
-                            if not first_symbol_table:
-                                first_symbol_table = symbol_table
-
-            if test_node is scope_limit:
-                # The ancestor scope/top-level Node has been reached and
-                # nothing has matched.
-                break
-
-            # Move on to the next ancestor.
-            test_node = test_node.parent
-
-        if possible_containers:
-            # No symbol found but there are one or more Containers from which
-            # it may be being brought into scope. Therefore create a generic
-            # Symbol with a deferred interface and add it to the most
-            # local SymbolTable with a wildcard import.
-            symbol = Symbol(name, interface=UnresolvedInterface(),
-                            visibility=visibility)
-            first_symbol_table.add(symbol)
-            return symbol
-
-        # All requested Nodes have been checked but there has been no
-        # match and there are no wildcard imports so raise an exception.
-        raise SymbolError(
-            "No Symbol found for name '{0}'.".format(name))
-
 
 # For automatic documentation generation
 # TODO #913 the 'colored' routine shouldn't be in this module.
