+++ src/psyclone/tests/psyir/nodes/node_test.py
@@ -48,8 +48,7 @@ from psyclone.psyir.nodes.node import (ChildrenList, Node,
 from psyclone.psyir.nodes import Schedule, Reference, Container, \
     Assignment, Return, Loop, Literal, Statement, node, KernelSchedule
 from psyclone.psyir.symbols import DataSymbol, SymbolError, \
-    INTEGER_TYPE, REAL_TYPE, SymbolTable, ContainerSymbol, \
-    UnresolvedInterface, ScalarType, DeferredType, Symbol
+    INTEGER_TYPE, REAL_TYPE, SymbolTable, ContainerSymbol
 from psyclone.psyGen import PSyFactory, OMPDoDirective, Kern
 from psyclone.errors import InternalError, GenerationError
 from psyclone.parse.algorithm import parse
@@ -668,144 +667,6 @@ def test_scope():
             "'Literal[value:'x', Scalar<INTEGER, UNDEFINED>]' as "
             "none of its ancestors are Container or Schedule nodes."
             in str(excinfo.value))
-
-
-def test_find_or_create_symbol():
-    '''Test that the find_or_create_symbol method in a Node instance
-    returns the associated symbol if there is one and raises an
-    exception if not. Also test for an incorrect scope argument.
-
-    '''
-    _, invoke = get_invoke("single_invoke_kern_with_global.f90",
-                           api="gocean1.0", idx=0)
-    sched = invoke.schedule
-    kernels = sched.walk(Kern)
-    kernel_schedule = kernels[0].get_kernel_schedule()
-    references = kernel_schedule.walk(Reference)
-
-    # Symbol in KernelSchedule SymbolTable
-    field_old = references[0]
-    assert field_old.name == "field_old"
-    assert isinstance(field_old.symbol, DataSymbol)
-    assert field_old.symbol in kernel_schedule.symbol_table.symbols
-
-    # Symbol in KernelSchedule SymbolTable with KernelSchedule scope
-    assert isinstance(field_old.find_or_create_symbol(
-        field_old.name, scope_limit=kernel_schedule), DataSymbol)
-    assert field_old.symbol.name == field_old.name
-
-    # Symbol in KernelSchedule SymbolTable with parent scope, so
-    # the symbol should not be found as we limit the scope to the
-    # immediate parent of the reference
-    with pytest.raises(SymbolError) as excinfo:
-        _ = field_old.find_or_create_symbol(field_old.name,
-                                            scope_limit=field_old.parent)
-    assert "No Symbol found for name 'field_old'." in str(excinfo.value)
-
-    # Symbol in Container SymbolTable
-    alpha = references[6]
-    assert alpha.name == "alpha"
-    assert isinstance(alpha.find_or_create_symbol(alpha.name), DataSymbol)
-    container = kernel_schedule.root
-    assert isinstance(container, Container)
-    assert (alpha.find_or_create_symbol(alpha.name) in
-            container.symbol_table.symbols)
-
-    # Symbol in Container SymbolTable with KernelSchedule scope, so
-    # the symbol should not be found as we limit the scope to the
-    # kernel so do not search the container symbol table.
-    with pytest.raises(SymbolError) as excinfo:
-        _ = alpha.find_or_create_symbol(alpha.name,
-                                        scope_limit=kernel_schedule)
-    assert "No Symbol found for name 'alpha'." in str(excinfo.value)
-
-    # Symbol in Container SymbolTable with Container scope
-    assert (alpha.find_or_create_symbol(
-        alpha.name, scope_limit=container).name == alpha.name)
-
-    # find_or_create_symbol method with invalid scope type
-    with pytest.raises(TypeError) as excinfo:
-        _ = alpha.find_or_create_symbol(alpha.name, scope_limit="hello")
-    assert ("The scope_limit argument 'hello' provided to the "
-            "find_or_create_symbol method, is not of type `Node`."
-            in str(excinfo.value))
-
-    # find_or_create_symbol method with invalid scope location
-    with pytest.raises(ValueError) as excinfo:
-        _ = alpha.find_or_create_symbol(alpha.name, scope_limit=alpha)
-    assert ("The scope_limit node 'Reference[name:'alpha']' provided to the "
-            "find_or_create_symbol method, is not an ancestor of this node "
-            "'Reference[name:'alpha']'." in str(excinfo.value))
-
-    # find_or_create_symbol method with invalid visibility
-    with pytest.raises(TypeError) as excinfo:
-        _ = alpha.find_or_create_symbol(alpha.name, visibility="hello")
-    assert ("visibility argument 'hello' provided to the find_or_create_symbol"
-            " method should be of `Symbol.Visibility` type but instead is: "
-            "'str'" in str(excinfo.value))
-
-    # With a valid visibility
-    sym = alpha.find_or_create_symbol("very_private",
-                                      visibility=Symbol.Visibility.PRIVATE)
-    assert sym.name == "very_private"
-    assert sym.visibility == Symbol.Visibility.PRIVATE
-    assert sym is container.symbol_table.lookup("very_private",
-                                                scope_limit=container)
-
-
-def test_find_or_create_new_symbol():
-    ''' Check that the Node.find_or_create_symbol() method creates new
-    symbols when appropriate. '''
-    # Create some suitable PSyIR from scratch
-    symbol_table = SymbolTable()
-    symbol_table.add(DataSymbol("tmp", REAL_TYPE))
-    kernel1 = KernelSchedule.create("mod_1", SymbolTable(), [])
-    container = Container.create("container_name", symbol_table,
-                                 [kernel1])
-    xvar = DataSymbol("x", REAL_TYPE)
-    xref = Reference(xvar)
-    assign = Assignment.create(xref, Literal("1.0", REAL_TYPE))
-    kernel1.addchild(assign)
-    assign.parent = kernel1
-    # We have no wildcard imports so there can be no symbol named 'undefined'
-    with pytest.raises(SymbolError) as err:
-        _ = assign.find_or_create_symbol("undefined")
-    assert "No Symbol found for name 'undefined'" in str(err.value)
-    # We should be able to find the 'tmp' symbol in the parent Container
-    sym = assign.find_or_create_symbol("tmp")
-    assert sym.datatype.intrinsic == ScalarType.Intrinsic.REAL
-    # Add a wildcard import to the SymbolTable of the KernelSchedule
-    new_container = ContainerSymbol("some_mod")
-    new_container.wildcard_import = True
-    kernel1.symbol_table.add(new_container)
-    # Symbol not in any container but we do have wildcard imports so we
-    # get a new symbol back
-    new_symbol = assign.find_or_create_symbol("undefined")
-    assert new_symbol.name == "undefined"
-    assert isinstance(new_symbol.interface, UnresolvedInterface)
-    assert type(new_symbol) == Symbol
-    assert "undefined" not in container.symbol_table
-    assert kernel1.symbol_table.lookup("undefined") is new_symbol
-
-
-def test_nemo_find_container_symbol(parser):
-    ''' Check that find_or_create_symbol() works for the NEMO API when the
-    searched-for symbol is declared in the parent module. '''
-    from fparser.common.readfortran import FortranFileReader
-    from psyclone.psyir.nodes import BinaryOperation
-    reader = FortranFileReader(
-        os.path.join(
-            os.path.dirname(os.path.dirname(os.path.dirname(
-                os.path.abspath(__file__)))),
-            "test_files", "gocean1p0", "kernel_with_global_mod.f90"))
-    prog = parser(reader)
-    psy = PSyFactory("nemo", distributed_memory=False).create(prog)
-    # Get a node from the schedule
-    bops = psy._invokes.invoke_list[0].schedule.walk(BinaryOperation)
-    # Use it as the starting point for the search
-    symbol = bops[0].find_or_create_symbol("alpha")
-    assert symbol.datatype.intrinsic == ScalarType.Intrinsic.REAL
-
 
 def test_children_validation():
     ''' Test that nodes are validated when inserted as children of other
